iommu: qcom: iommu-v1: Add SECURE_VMID, CONTEXT_BANK, etc attributes
Add handling for the DOMAIN_ATTR_{SECURE_VMID, CONTEXT_BANK, ATOMIC,
PT_BASE_ADDR} IOMMU domain attributes.

Some of these attributes are basically NULL and just a way to
"fail gracefully" as some features are not supported on MMU-500
units and/or QC secure IOMMU implementation: for example, the
secure VMID setting can be done only once, and for MMU-500
equipped hardware, this is being already done early in the
bootchain.
About context bank attribute, it's unusable because this driver
is actually using static context banks, also defined in the
early bootchain, which are LOCKED and SECURELY allocated, so
they cannot be "just changed" on-the-fly.
Lastly, the atomic attribute is just useless, as our (un)mapping
is ALWAYS atomic by design, so nothing can be changed here.

In get_attr, I have introduced the PT_BASE_ADDR attribute
handling, which is supposedly deprecated, as it isn't currently
used by any driver (as of k4.4) and it is scheduled for removal
in the future.