iommu: arm-smmu-legacy: FIXME: HACK! Do not reset SMMU device on attach
Whenever we try to reset a secure context, the device will hard
lockup and will reboot becauses of a breach in the security.

Hack to not reset any context: this is bad, but it works for now
and there's no time to fix it properly, unfortunately.
It will be fixed as soon as possible to restore clean operation.